20 - LT - Move husky views to central location
