+++ src/app/views/appviews/huskyViews/huskyHappiest.component.html
