+++ src/app/views/appviews/huskyViews/huskyHappiest.component.ts
@@ -3,7 +3,7 @@ import { Component, OnDestroy, OnInit } from "@angular/core";
 @Component({
   selector: 'husky-happiest',
   templateUrl: './huskyHappiest.component.html',
-  styleUrls: ['component.css']
+  styleUrls: ['../component.css']
 })
 
 export class HuskyHappiestComponent implements OnDestroy, OnInit {
