+++ src/app/views/appviews/huskyViews/huskySad.component.ts
@@ -5,7 +5,7 @@ import { Component, OnDestroy, OnInit } from "@angular/core";
   template: `
     <div class="row husky-sad justify-content-md-center"></div>
   `,
-  styleUrls: ['husky.component.css']
+  styleUrls: ['./husky.component.css']
 })
 
 export class HuskySadComponent implements OnDestroy, OnInit {
